+++ apps/laboratory/src/components/Ethers/EthersTransactionTest.tsx
@@ -1,7 +1,7 @@
 import { Button, useToast, Stack, Link, Text, Spacer } from '@chakra-ui/react'
 import { useWeb3ModalAccount, useWeb3ModalProvider } from '@web3modal/ethers/react'
 import { BrowserProvider, JsonRpcSigner, ethers } from 'ethers'
-import { sepolia } from '../../utils/ChainsUtil'
+import { sepolia, optimism } from '../../utils/ChainsUtil'
 import { useState } from 'react'
 import { vitalikEthAddress } from '../../utils/DataUtil'
 
@@ -21,9 +21,7 @@ export function EthersTransactionTest() {
       const signer = new JsonRpcSigner(provider, address)
       const tx = await signer.sendTransaction({
         to: vitalikEthAddress,
-        value: ethers.parseUnits('0.0001', 'gwei'),
-        maxFeePerGas: ethers.parseUnits('200', 'gwei'),
-        maxPriorityFeePerGas: ethers.parseUnits('200', 'gwei')
+        value: ethers.parseUnits('0.0001', 'gwei')
       })
       toast({ title: 'Succcess', description: tx.hash, status: 'success', isClosable: true })
     } catch {
@@ -38,7 +36,9 @@ export function EthersTransactionTest() {
     }
   }
 
-  return chainId === sepolia.chainId && address ? (
+  const allowedChains = [sepolia.chainId, optimism.chainId]
+
+  return allowedChains.includes(Number(chainId)) && address ? (
     <Stack direction={['column', 'column', 'row']}>
       <Button
         data-test-id="sign-transaction-button"
@@ -64,7 +64,7 @@ export function EthersTransactionTest() {
     </Stack>
   ) : (
     <Text fontSize="md" color="yellow">
-      Switch to Sepolia Ethereum Testnet to test this feature
+      Switch to Sepolia or OP to test this feature
     </Text>
   )
 }