+++ packages/common/index.ts
@@ -1,4 +1,6 @@
 // -- Utils -------------------------------------------------------------------
 export { DateUtil } from './src/utils/DateUtil.js'
+export { NetworkUtil } from './src/utils/NetworkUtil.js'
+export { NumberUtil } from './src/utils/NumberUtil.js'
 
 export type * from './src/utils/TypeUtil.js'